# Commit Title: Refactor StressNgAdapter for Direct LoggerPort Injection
## Commit Description:

This commit enhances the `StressNgAdapter` structure in the project. The primary change includes the removal of the '`self`' reference for the `logger` in the methods of the `StressNgAdapter` and instead directly injecting the `logger` from `LoggerPort`.

The following methods were updated:

1. `decide_stress_ng_arch()`
2. `prepare_stress_ng_binary()`
3. `execute_stress_ng_command()`
4. `remove_stress_ng_binary()`

Each of these methods now takes in '`logger`' directly: